[ios, macos] Rename MGLMetricsManager to MGLSDKMetricsManager.
