+++ platform/ios/sdk-files.json
@@ -65,7 +65,7 @@
         "platform/darwin/src/MGLComputedShapeSource.mm",
         "platform/darwin/src/MGLDistanceFormatter.m",
         "platform/ios/vendor/mapbox-events-ios/vendor/TrustKit/TSKPinningValidatorResult.m",
-        "platform/darwin/src/MGLMetricsManager.m",
+        "platform/darwin/src/MGLSDKMetricsManager.m",
         "platform/ios/vendor/mapbox-events-ios/MapboxMobileEvents/MMETrustKitProvider.m",
         "platform/ios/vendor/mapbox-events-ios/MapboxMobileEvents/MMEAPIClient.m",
         "platform/ios/src/MGLMapView.mm",
@@ -163,7 +163,7 @@
         "MGLVectorStyleLayer.h": "platform/darwin/src/MGLVectorStyleLayer.h",
         "MGLNetworkConfiguration.h": "platform/darwin/src/MGLNetworkConfiguration.h",
         "MGLComputedShapeSource.h": "platform/darwin/src/MGLComputedShapeSource.h",
-        "MGLMetricsManager.h": "platform/darwin/src/MGLMetricsManager.h",
+        "MGLSDKMetricsManager.h": "platform/darwin/src/MGLSDKMetricsManager.h",
         "MGLLoggingConfiguration.h": "platform/darwin/src/MGLLoggingConfiguration.h",
         "MGLLocationManager.h": "platform/darwin/src/MGLLocationManager.h",
         "MGLRasterDEMSource.h": "platform/darwin/src/MGLRasterDEMSource.h",
@@ -225,7 +225,7 @@
         "MGLStyle_Private.h": "platform/darwin/src/MGLStyle_Private.h",
         "MGLHeatmapStyleLayer_Private.h": "platform/darwin/src/MGLHeatmapStyleLayer_Private.h",
         "NSPredicate+MGLPrivateAdditions.h": "platform/darwin/src/NSPredicate+MGLPrivateAdditions.h",
-        "MGLMetricsManager_Private.h": "platform/darwin/src/MGLMetricsManager_Private.h",
+        "MGLSDKMetricsManager_Private.h": "platform/darwin/src/MGLSDKMetricsManager_Private.h",
         "MGLNetworkConfiguration_Private.h": "platform/darwin/src/MGLNetworkConfiguration_Private.h",
         "NSDate+MGLAdditions.h": "platform/darwin/src/NSDate+MGLAdditions.h",
         "MGLCompactCalloutView.h": "platform/ios/src/MGLCompactCalloutView.h",
