+++ platform/ios/src/Mapbox.h
@@ -73,4 +73,4 @@ FOUNDATION_EXPORT MGL_EXPORT const unsigned char MapboxVersionString[];
 #import "MGLLoggingConfiguration.h"
 #import "MGLNetworkConfiguration.h"
 #import "MGLAttributedExpression.h"
-#import "MGLMetricsManager.h"
+#import "MGLSDKMetricsManager.h"
