+++ platform/macos/sdk-files.json
@@ -6,7 +6,7 @@
         "platform/macos/src/NSImage+MGLAdditions.mm",
         "platform/darwin/src/MGLHillshadeStyleLayer.mm",
         "platform/darwin/src/MGLPointAnnotation.mm",
-        "platform/darwin/src/MGLMetricsManager.m",
+        "platform/darwin/src/MGLSDKMetricsManager.m",
         "platform/darwin/src/NSBundle+MGLAdditions.m",
         "platform/macos/src/MGLMapView.mm",
         "platform/darwin/src/NSArray+MGLAdditions.mm",
@@ -93,7 +93,7 @@
         "MGLMapView.h": "platform/macos/src/MGLMapView.h",
         "MGLOfflineStorage.h": "platform/darwin/src/MGLOfflineStorage.h",
         "MGLStyleValue.h": "platform/darwin/src/MGLStyleValue.h",
-        "MGLMetricsManager.h": "platform/darwin/src/MGLMetricsManager.h",
+        "MGLSDKMetricsManager.h": "platform/darwin/src/MGLSDKMetricsManager.h",
         "MGLMultiPoint.h": "platform/darwin/src/MGLMultiPoint.h",
         "MGLShapeCollection.h": "platform/darwin/src/MGLShapeCollection.h",
         "MGLPointAnnotation.h": "platform/darwin/src/MGLPointAnnotation.h",
@@ -173,7 +173,7 @@
         "NSException+MGLAdditions.h": "platform/darwin/src/NSException+MGLAdditions.h",
         "MGLTileSource_Private.h": "platform/darwin/src/MGLTileSource_Private.h",
         "NSExpression+MGLPrivateAdditions.h": "platform/darwin/src/NSExpression+MGLPrivateAdditions.h",
-        "MGLMetricsManager_Private.h": "platform/darwin/src/MGLMetricsManager_Private.h",
+        "MGLSDKMetricsManager_Private.h": "platform/darwin/src/MGLSDKMetricsManager_Private.h",
         "MGLNetworkConfiguration_Private.h": "platform/darwin/src/MGLNetworkConfiguration_Private.h",
         "NSCompoundPredicate+MGLAdditions.h": "platform/darwin/src/NSCompoundPredicate+MGLAdditions.h",
         "MGLSymbolStyleLayer_Private.h": "platform/darwin/src/MGLSymbolStyleLayer_Private.h",
