+++ platform/macos/src/Mapbox.h
@@ -68,4 +68,4 @@ FOUNDATION_EXPORT MGL_EXPORT const unsigned char MapboxVersionString[];
 #import "MGLLoggingConfiguration.h"
 #import "MGLNetworkConfiguration.h"
 #import "MGLAttributedExpression.h"
-#import "MGLMetricsManager.h"
+#import "MGLSDKMetricsManager.h"
